Controller.cpp: Respect NoteOn with Velocity=0 as NoteOff
A lot of midi keyboards, (Novaton LaunchKey MK3 mini for example)
send a NoteOn with Velocity=0 as a NoteOff. Current Drumboy firmware
seems to expect a NoteOff, but as a NoteOff never occurs, the note just
keeps playing.

This is an untested patch (need the STLink etc) but it should in theory
fix things.

Signed-off-by: Eilís 'pidge' Ní Fhlannagáin <pidge@baylibre.com>